+++ changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Testing
     * Unit tests for notes field rules of new borrowing request.
     * New borrowing request.
+    * End borrowing request.
 
 ## [0.12.0] - 2018-10-06